Restructured Pages
Put components inside the route directories
