+++ app/about/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Base from "../components/layouts/base";
-import ProfileHeader from "../components/pages/header";
+import ProfileHeader from "../components/ui/header";
 import ContainerLayout from "../components/layouts/container";
 import InProgressImage from "../../public/aboutme-option10.webp";
 import Image from "next/image";
@@ -10,7 +10,7 @@ const page = () => {
     <Base>
       <ContainerLayout className="flex flex-col space-y-4">
         <h1 className="text-3xl font-semibold tracking-tight text-gray-12 md:text-4xl">
-          Blog
+          About Me
         </h1>
         <div className="text-gray-11">Work in progress.</div>
         <Image
